+++ JobMaster.Api/ApiModels/ApiBucketModel.cs
@@ -8,6 +8,7 @@ public class ApiBucketModel : ApiClusterBaseModel
     public string Id { get; set; } = string.Empty;
     public string Name { get; set; } = string.Empty;
     public string AgentConnectionId { get; set; } = null!;
+    public string AgentConnectionName { get; set; } = null!;
     public string? AgentWorkerId { get; set; }
     public string RepositoryTypeId { get; set; } = string.Empty;
     public JobMasterPriority Priority { get; set; }
@@ -22,6 +23,7 @@ public class ApiBucketModel : ApiClusterBaseModel
         return new ApiBucketModel()
         {
             AgentConnectionId = model.AgentConnectionId.IdValue,
+            AgentConnectionName = model.AgentConnectionId.Name,
             AgentWorkerId = model.AgentWorkerId,
             Priority = model.Priority,
             Status = model.Status,