+++ JobMaster.SampleWeb/Program.cs
@@ -14,7 +14,85 @@ using Microsoft.OpenApi.Models;
 using Serilog;
 
 var builder = WebApplication.CreateBuilder(args);
-
+//
+// builder.Services.AddJobMasterCluster(config =>
+// {
+//     config.ClusterId("Cluster-1")
+//           .ClusterTransientThreshold(TimeSpan.FromMinutes(1))
+//           .DebugJsonlFileLogger("/home/hugo/logs/Cluster-1.log")
+//           .ClusterMode(ClusterMode.Active);
+//
+//     // Master database (must be SQL)
+//     config.UsePostgresForMaster("Host=localhost;Port=5432;Database=jobmaster;Username=postgres;Password=postgres;Maximum Pool Size=300");
+//     
+//     // config.AddAgentConnectionConfig("Pg-1")
+//     //       .UsePostgresForAgent("Host=localhost;Port=5432;Database=agent_pg1;Username=postgres;Password=postgres");
+//     //
+//     // config.AddAgentConnectionConfig("My-1")
+//     //       .UseMySqlForAgent("Server=localhost;Port=3306;Database=agent_my1;User ID=root;Password=root;");
+//     //
+//     // config.AddAgentConnectionConfig("Sql-1")
+//     //     .UseSqlServerForAgent("Server=localhost,1433;Initial Catalog=agent_sql1;User Id=sa;Password=Passw0rd!;Encrypt=False;TrustServerCertificate=True;");
+//
+//     
+//     config.AddAgentConnectionConfig("Nats-1")
+//           .UseNatsJetStream("nats://jmuser:jmpass@localhost:4222");
+//
+//     var isConsumer = Environment.GetEnvironmentVariable("CONSUMER")?.ToUpperInvariant() == "TRUE";
+//     if (isConsumer)
+//     {
+//        config.AddWorker()
+//            .AgentConnName("Nats-1")
+//            .BucketQtyConfig(JobMasterPriority.Medium, 1)
+//            .WorkerBatchSize(1000)
+//            .SetWorkerMode(AgentWorkerMode.Full);
+//
+//        config.AddWorker()
+//            .AgentConnName("Nats-1")
+//            .BucketQtyConfig(JobMasterPriority.Medium, 1)
+//            .WorkerBatchSize(1000)
+//            .SetWorkerMode(AgentWorkerMode.Drain);
+//     }
+// });//
+// builder.Services.AddJobMasterCluster(config =>
+// {
+//     config.ClusterId("Cluster-1")
+//           .ClusterTransientThreshold(TimeSpan.FromMinutes(1))
+//           .DebugJsonlFileLogger("/home/hugo/logs/Cluster-1.log")
+//           .ClusterMode(ClusterMode.Active);
+//
+//     // Master database (must be SQL)
+//     config.UsePostgresForMaster("Host=localhost;Port=5432;Database=jobmaster;Username=postgres;Password=postgres;Maximum Pool Size=300");
+//     
+//     // config.AddAgentConnectionConfig("Pg-1")
+//     //       .UsePostgresForAgent("Host=localhost;Port=5432;Database=agent_pg1;Username=postgres;Password=postgres");
+//     //
+//     // config.AddAgentConnectionConfig("My-1")
+//     //       .UseMySqlForAgent("Server=localhost;Port=3306;Database=agent_my1;User ID=root;Password=root;");
+//     //
+//     // config.AddAgentConnectionConfig("Sql-1")
+//     //     .UseSqlServerForAgent("Server=localhost,1433;Initial Catalog=agent_sql1;User Id=sa;Password=Passw0rd!;Encrypt=False;TrustServerCertificate=True;");
+//
+//     
+//     config.AddAgentConnectionConfig("Nats-1")
+//           .UseNatsJetStream("nats://jmuser:jmpass@localhost:4222");
+//
+//     var isConsumer = Environment.GetEnvironmentVariable("CONSUMER")?.ToUpperInvariant() == "TRUE";
+//     if (isConsumer)
+//     {
+//        config.AddWorker()
+//            .AgentConnName("Nats-1")
+//            .BucketQtyConfig(JobMasterPriority.Medium, 1)
+//            .WorkerBatchSize(1000)
+//            .SetWorkerMode(AgentWorkerMode.Full);
+//
+//        config.AddWorker()
+//            .AgentConnName("Nats-1")
+//            .BucketQtyConfig(JobMasterPriority.Medium, 1)
+//            .WorkerBatchSize(1000)
+//            .SetWorkerMode(AgentWorkerMode.Drain);
+//     }
+// });//
 builder.Services.AddJobMasterCluster(config =>
 {
     config.ClusterId("Cluster-1")
@@ -53,13 +131,68 @@ builder.Services.AddJobMasterCluster(config =>
            .WorkerBatchSize(1000)
            .SetWorkerMode(AgentWorkerMode.Drain);
     }
-});
+});//
+// builder.Services.AddJobMasterCluster(config =>
+// {
+//     config.ClusterId("Cluster-1")
+//           .ClusterTransientThreshold(TimeSpan.FromMinutes(1))
+//           .DebugJsonlFileLogger("/home/hugo/logs/Cluster-1.log")
+//           .ClusterMode(ClusterMode.Active);
+//
+//     // Master database (must be SQL)
+//     config.UsePostgresForMaster("Host=localhost;Port=5432;Database=jobmaster;Username=postgres;Password=postgres;Maximum Pool Size=300");
+//     
+//     // config.AddAgentConnectionConfig("Pg-1")
+//     //       .UsePostgresForAgent("Host=localhost;Port=5432;Database=agent_pg1;Username=postgres;Password=postgres");
+//     //
+//     // config.AddAgentConnectionConfig("My-1")
+//     //       .UseMySqlForAgent("Server=localhost;Port=3306;Database=agent_my1;User ID=root;Password=root;");
+//     //
+//     // config.AddAgentConnectionConfig("Sql-1")
+//     //     .UseSqlServerForAgent("Server=localhost,1433;Initial Catalog=agent_sql1;User Id=sa;Password=Passw0rd!;Encrypt=False;TrustServerCertificate=True;");
+//
+//     
+//     config.AddAgentConnectionConfig("Nats-1")
+//           .UseNatsJetStream("nats://jmuser:jmpass@localhost:4222");
+//
+//     var isConsumer = Environment.GetEnvironmentVariable("CONSUMER")?.ToUpperInvariant() == "TRUE";
+//     if (isConsumer)
+//     {
+//        config.AddWorker()
+//            .AgentConnName("Nats-1")
+//            .BucketQtyConfig(JobMasterPriority.Medium, 1)
+//            .WorkerBatchSize(1000)
+//            .SetWorkerMode(AgentWorkerMode.Full);
+//
+//        config.AddWorker()
+//            .AgentConnName("Nats-1")
+//            .BucketQtyConfig(JobMasterPriority.Medium, 1)
+//            .WorkerBatchSize(1000)
+//            .SetWorkerMode(AgentWorkerMode.Drain);
+//     }
+// });
 
-builder.Services.AddJobMasterCluster(c =>
-        c.UseClusterStandalone().
-            ClusterId("Cluster-Standalone-1").
-        .UsePostgresForMaster("Host=localhost;Port=5432;Database=jobmaster_standalone;Username=postgres;Password=postgres;Maximum Pool Size=300")
-    );
+builder.Services.AddJobMasterCluster(c => {
+        
+    c.UseStandaloneCluster().ClusterId("Cluster-Standalone-1")
+        .UsePostgres("Host=localhost;Port=5432;Database=jobmaster_standalone;Username=postgres;Password=postgres;Maximum Pool Size=300")
+        .SetAsDefault()
+        .AddWorker();
+    
+        // c.ClusterId("Cluster-Standalone-1")
+        //     .UsePostgresForMaster("Host=localhost;Port=5432;Database=jobmaster_standalone;Username=postgres;Password=postgres;Maximum Pool Size=300")
+        //     .SetAsDefault();
+        //
+        //
+        // c.AddAgentConnectionConfig("Nats-1")
+        //     .UseNatsJetStream("nats://jmuser:jmpass@localhost:4222");
+        //
+        // c.AddWorker()
+        //     .AgentConnName("Nats-1")
+        //     .BucketQtyConfig(JobMasterPriority.Medium, 1)
+        //     .WorkerBatchSize(1000)
+        //     .SetWorkerMode(AgentWorkerMode.Full);
+});
 
 builder.Services.UseJobMasterApi(o =>
 {
@@ -70,6 +203,9 @@ builder.Services.UseJobMasterApi(o =>
     o.UseApiKeyAuth().AddApiKey("my-api-key", "admin-key");
     o.UseUserPwdAuth().AddUserPwd("john", "pwd#123");
 });
+
+
+
 builder.Services.AddEndpointsApiExplorer();
 
 builder.Services.AddSwaggerGen();
@@ -84,7 +220,7 @@ builder.Services.AddSerilog();
 
 var app = builder.Build();
 
-app.StartJobMasterApi();
+app.MapJobMasterApi();
 
 await app.Services.StartJobMasterRuntimeAsync();
 
@@ -94,7 +230,7 @@ app.UseSwaggerUI();
 app.UseHttpsRedirection();
 
 
-app.MapPost("/schedule-job", async (int qty, string? lane, TimeSpan? delay, IJobMasterScheduler jobScheduler) =>
+app.MapPost("/schedule-job", async(int qty, string ? lane, string? clusterId, TimeSpan? delay, IJobMasterScheduler jobScheduler) =>
 {
     if (string.IsNullOrWhiteSpace(lane)) lane = null;
 
@@ -104,9 +240,9 @@ app.MapPost("/schedule-job", async (int qty, string? lane, TimeSpan? delay, IJob
     {
         var data = WriteableMessageData.New().SetStringValue("Name", Faker.Name.FullName());
         if (delay.HasValue)
-            tasks.Add(jobScheduler.OnceAfterAsync<HelloJobHandler>(delay.Value, data, metadata: meta, workerLane: lane));
+            tasks.Add(jobScheduler.OnceAfterAsync<HelloJobHandler>(delay.Value, data, metadata: meta, workerLane: lane, clusterId: clusterId));
         else
-            tasks.Add(jobScheduler.OnceNowAsync<HelloJobHandler>(data, metadata: meta, workerLane: lane));
+            tasks.Add(jobScheduler.OnceNowAsync<HelloJobHandler>(data, metadata: meta, workerLane: lane, clusterId: clusterId));
     }
 
     await Task.WhenAll(tasks);